feat(analyzers): restore DXA010-DXA080 suite from archive
- Restored analyzers, infrastructure, and ResultFlow from archive/analyzers
- Excluded docs/ and READMEs to avoid Windows case conflicts
- Tests and rulesets unchanged
