+++ src/Dx.Domain.Analyzers/Infrastructure/Facades/IDxFacadeResolver.cs
@@ -14,7 +14,6 @@ using System.Collections.Generic;
 using System.Linq;
 
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.Diagnostics;
 
 namespace Dx.Domain.Analyzers.Infrastructure.Facades
 {
@@ -41,81 +40,5 @@ namespace Dx.Domain.Analyzers.Infrastructure.Facades
         /// <param name="type">The candidate result type.</param>
         /// <returns>The first matching factory method, or <see langword="null"/> if none is found.</returns>
         IMethodSymbol? FindFacadeFactoryForType(ITypeSymbol type);
-    }
-
-    /// <summary>
-    /// Default implementation that reflects over the <c>Dx</c> root facade in <c>Dx.Domain</c>
-    /// (or a configured alternative) and collects public static factory methods.
-    /// </summary>
-    public sealed class DxFacadeResolver : IDxFacadeResolver
-    {
-        private readonly HashSet<IMethodSymbol> _methods =
-            new(SymbolEqualityComparer.Default);
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="DxFacadeResolver"/> class.
-        /// </summary>
-        /// <param name="compilation">The compilation used to resolve the facade type.</param>
-        /// <param name="config">Analyzer configuration options used to locate the root facade.</param>
-        public DxFacadeResolver(Compilation compilation, AnalyzerConfigOptionsProvider config)
-        {
-            // Allow the root facade type to be overridden via EditorConfig, falling back to the
-            // fully qualified metadata name of the canonical root facade type "Dx.Dx".
-            var rootTypeName = GetRootFacadeTypeName(config) ?? "Dx.Dx";
-
-            // First, try to bind the configured or default root facade type directly.
-            var dx = compilation.GetTypeByMetadataName(rootTypeName);
-
-            // As a resilience measure for older configurations that might still use the short
-            // type name "Dx", fall back to that if the fully qualified lookup fails.
-            dx ??= compilation.GetTypeByMetadataName("Dx");
-
-            if (dx == null)
-                return;
-
-            // We treat all public nested types of the root facade as logical namespaces
-            // containing public static factory methods.
-            foreach (var nested in dx.GetTypeMembers())
-            {
-                if (nested.DeclaredAccessibility != Accessibility.Public)
-                    continue;
-
-                foreach (var method in nested.GetMembers().OfType<IMethodSymbol>())
-                {
-                    if (method.DeclaredAccessibility != Accessibility.Public ||
-                        !method.IsStatic)
-                    {
-                        continue;
-                    }
-
-                    // Future extension point: filter to methods returning Dx.Domain kernel types
-                    // or Result<T>. For now, record all public static methods and let
-                    // individual rules apply additional constraints.
-                    _methods.Add(method);
-                }
-            }
-        }
-
-        /// <inheritdoc />
-        public IReadOnlyCollection<IMethodSymbol> FacadeFactories => _methods;
-
-        /// <inheritdoc />
-        public bool IsDxFacadeFactory(IMethodSymbol method) =>
-            _methods.Contains(method);
-
-        /// <inheritdoc />
-        public IMethodSymbol? FindFacadeFactoryForType(ITypeSymbol type) =>
-            _methods.FirstOrDefault(m =>
-                SymbolEqualityComparer.Default.Equals(m.ReturnType, type));
-
-        private static string? GetRootFacadeTypeName(AnalyzerConfigOptionsProvider config)
-        {
-            // Config key aligned with docs: "dx_facade_root". When not present, callers fall
-            // back to the conventional root type name "Dx.Dx".
-            var global = config.GlobalOptions;
-            return global.TryGetValue("dx_facade_root", out var value) && !string.IsNullOrWhiteSpace(value)
-                ? value.Trim()
-                : null;
-        }
     }
 }
